separate out core and example dir
