+++ example/src/App.jsx
@@ -1,7 +1,6 @@
-import { h } from "./core/h";
-import { useState } from "./core/hooks/useState";
+import { h, useState } from "../../src/public-api";
 
-export const App = () => {
+const App = () => {
     console.log('running/ building app')
     const [count, setCount] = useState(0);
     const [count1, setCount1] = useState(0);
@@ -15,12 +14,11 @@ export const App = () => {
         text
     );
 
-    // const button = createElement('button', {}, `Count`)
     const button = h('button', {onClick: () => setCount(count + 1)}, `Count ${count}`);
-    const button1 = h('button', {onClick: () => setCount1(count + 1)}, `Count ${count1}`);
-    const div = h('div', {}, button, button1);
+    const button1 = h('button', {onClick: () => setCount1(count1 + 1)}, `Count ${count1}`);
+    const div = h('div', {}, button, button1, title);
 
-    // return h('div', {}, title, button)
-    // return createElement('div', {}, button)
-    return div;
+    return div
 }
+
+export default App;
