+++ example/src/main.js
@@ -1,6 +1,5 @@
-import './style.css'
-import { render } from './core/render';
-import { App } from './App';
+import { render } from '../../src/public-api'
+import App from './App';
 
 const root = document.getElementById("root");
 console.log('in main.js')
