+++ package.json
@@ -4,7 +4,7 @@
   "version": "1.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite example",
     "build": "vite build",
     "preview": "vite preview"
   },
