+++ src/public-api.js
@@ -1,0 +1,5 @@
+import h from "./core/h";
+import render from "./core/render";
+import { useState } from "./core/hooks/useState";
+
+export { h, render, useState };
